Modify handler methods and path
