+++ JSONdata/labels.json
@@ -1,30 +1,42 @@
 [
   {
-    "label_id" : 814,
+    "label_id" : 818,
     "title" : "New Physics",
     "color" : "Blue",
     "Items" : [
       {
-        "cover_state" : "P.N Okeke",
-        "publisher" : "good",
+        "cover_state" : "good",
+        "publisher" : "P.N Okeke",
         "publish_date" : "1999-04-28"
       },
       {
-        "cover_state" : "bad",
-        "publisher" : "Elizabeth Meyer",
+        "cover_state" : "Elizabeth Meyer",
+        "publisher" : "bad",
         "publish_date" : "2004-08-25"
       }
     ]
   },
   {
-    "label_id" : 484,
+    "label_id" : 660,
     "title" : "Chemistry",
     "color" : "brown",
     "Items" : [
       {
-        "cover_state" : "excellent",
-        "publisher" : "Ababio Johnson",
+        "cover_state" : "Ababio Johnson",
+        "publisher" : "excellent",
         "publish_date" : "2019-07-18"
+      }
+    ]
+  },
+  {
+    "label_id" : 758,
+    "title" : "Grownish",
+    "color" : "Magenta",
+    "Items" : [
+      {
+        "cover_state" : "good",
+        "publisher" : "Alice Stewart",
+        "publish_date" : "1975-04-29"
       }
     ]
   }
