+++ handler/helper_methods.rb
@@ -1,0 +1,23 @@
+require 'json'
+
+module HelperFunctions
+  # WRITER
+  def write_json(array, file_path)
+    opts = {
+      array_nl: "\n",
+      object_nl: "\n",
+      indent: '  ',
+      space_before: ' ',
+      space: ' '
+    }
+    File.write(file_path, JSON.generate(array, opts))
+  end
+
+  # READER
+  def read_json(file_path)
+    return unless File.exist?(file_path)
+
+    file = File.read(file_path)
+    JSON.parse(file)
+  end
+end
